requirements.txt
absl-py==2.1.0
addict==2.4.0
albumentations @ git+https://github.com/albumentations-team/albumentations@3c510e1422f6663ccd5026950af2d9ea85b5302a
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
antlr4-python3-runtime==4.9.3
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
attrs==23.2.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
batchgenerators==0.25
black==24.2.0
blessed==1.20.0
cachetools==5.3.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
colorlog==6.8.2
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1691044910542/work
contextlib2==21.6.0
contourpy==1.2.0
cosine-annealing-warmup @ git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup@12d03c07553aedd3d9e9155e2b3e31ce8c64081a
crcmod==1.7
cryptography==41.0.7
cycler==0.12.1
Cython==3.0.2
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1695534290310/work
decorator==4.4.2
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@a59f05630a8f205756064244bf5beb8661f96180
efficientnet-pytorch==0.7.1
einops==0.7.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1692026125334/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
fairscale==0.4.13
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
ftfy==6.2.0
future==1.0.0
fvcore==0.1.5.post20221221
google-auth==2.22.0
google-auth-oauthlib==1.0.0
googledrivedownloader==0.4
gpustat==1.1.1
grad-cam==1.5.0
grpcio==1.62.0
h5py==3.10.0
hausdorff==0.2.6
huggingface-hub==0.21.4
hydra-core==1.3.2
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
imgaug==0.4.0
importlib_metadata==7.0.2
importlib_resources==6.1.3
iopath==0.1.9
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1698244021190/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1696264049390/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.3.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.4
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1698231991356/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1698673647019/work
kiwisolver==1.4.5
kornia==0.7.1
lazy_loader==0.3
lightning-utilities==0.10.1
linecache2==1.0.0
llvmlite==0.42.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
MedPy==0.4.0
ml-collections==0.1.1
mmcv==2.1.0
mmcv-full==1.7.0
mmdet==3.3.0
mmengine==0.10.2
model-index==0.1.11
monai==1.3.0
moviepy==1.0.3
mpmath==1.3.0
MultiScaleDeformableAttention==1.0
munch==4.0.0
mypy-extensions==1.0.0
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1697083700168/work
networkx==3.2.1
nibabel==5.2.1
nltk==3.8.1
nptyping==2.5.0
numba==0.59.0
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1695290862901/work/dist/numpy-1.26.0-cp310-cp310-linux_x86_64.whl#sha256=44509c98ccedaff13cf312f80a8e392a35d5f649bdf63f1b7e705fdfdc621c6d
nvidia-ml-py==12.535.133
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.24.0
opencv-python==4.9.0.80
opencv-python-headless==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.34
ordered-set==4.1.0
oss2==2.17.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==2.1.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathspec==0.12.1
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.2.0
pkuseg==0.0.25
platformdirs==4.2.0
portalocker==2.8.2
pretrainedmodels==0.7.4
proglog==0.1.10
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1688565951714/work
protobuf==4.25.3
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1695367094274/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
pydensecrf @ file:///home/conda/feedstock_root/build_artifacts/pydensecrf_1667219279105/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1691408637400/work
PYMIC==0.4.0
pynrrd==1.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1698062401223/work
qudida==0.0.4
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
segmentation-models-pytorch @ git+https://github.com/qubvel/segmentation_models.pytorch@6db76a1106426ac5b55f39fba68168f3bccae7f8
sentencepiece==0.2.0
shapely==2.0.3
SimpleITK==2.3.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
submitit==1.5.1
supervision==0.6.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2024.2.12
timm==0.9.2
tokenizers==0.15.2
tomli==2.0.1
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchmetrics==1.3.1
torchnet==0.0.4
torchvision==0.14.1+cu116
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1695373560918/work
tqdm==4.66.2
traceback2==1.4.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1698671135544/work
transformers==4.39.1
triton==2.2.0
ttach==0.0.3
typing_extensions==4.10.0
tzdata==2023.3
unittest2==1.1.0
urllib3==2.2.1
visdom==0.2.4
wcwidth==0.2.13
websocket-client==1.6.3
Werkzeug==3.0.1
yacs==0.1.8
yapf==0.40.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
